+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Master](https://github.com/appwise-labs/AppwiseCore)
 
+## [0.10.7](https://github.com/appwise-labs/AppwiseCore/releases/tag/0.10.7)
+
 ### New Features
 
 * Core: split off the application delegate and services into a separate subspec `UIApplication`, so that the `Core` subspec can be used in application extensions.